Move to a new `OpenDreamPackageTool` project Ran into file locking issues with OpenDreamPackaging as a dependency
